+++ lib/djv/App/FilesTool.h
@@ -76,6 +76,7 @@ namespace djv
             void _aUpdate(const std::shared_ptr<models::FilesModelItem>&);
             void _bUpdate(const std::vector<std::shared_ptr<models::FilesModelItem> >&);
             void _layersUpdate(const std::vector<int>&);
+            void _mediaReferenceKeysUpdate();
             void _compareUpdate(const tl::CompareOptions&);
 
             FTK_PRIVATE();